+++ programs/encore/src/instructions/mod.rs
@@ -7,6 +7,7 @@ pub mod listing_close;
 pub mod listing_complete;
 pub mod listing_create;
 pub mod listing_release;
+pub mod listing_seller_cancel_claim;
 pub mod ticket_mint;
 pub mod ticket_transfer;
 
@@ -19,5 +20,6 @@ pub use listing_close::*;
 pub use listing_complete::*;
 pub use listing_create::*;
 pub use listing_release::*;
+pub use listing_seller_cancel_claim::*;
 pub use ticket_mint::*;
 pub use ticket_transfer::*;